hephaes/pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hephaes"
dynamic = ["version"]
description = "ROS log processing and dataset conversion"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  {name = "Daniel Yoo"},
  {name = "Fleming Yip"},
]
keywords = ["robotics", "ros", "rosbag", "mcap", "parquet", "tfrecord", "datasets"]
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "numpy>=1.26.0,<2.0",
  "rosbags>=0.9.0",
  "PyYAML>=6.0",
  "pydantic>=2.6.0,<3.0.0",
  "pyarrow>=16.0.0",
  "orjson>=3.9.0",
]

[project.urls]
Homepage = "https://hephaes.ai"

[project.optional-dependencies]
dev = [
  "build>=1.2.2,<2.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.0.0,<9.0.0",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "hephaes._version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
